requirements.txt
Django>=5.0,<6.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
reportlab>=4.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
gunicorn>=22.0.0
